+++ server/src/v1/controllers/section.js
@@ -2,8 +2,17 @@ const Section = require('../models/section')
 const Task = require('../models/task')
 
 exports.create = async (req, res) => {
+  const { user } = req
   const { boardId } = req.params
+
   try {
+    if (!user) {
+      throw new Error('User not found');
+    }
+    if (user.role !== 'Админ') {
+      return res.status(403).json({ error: 'У вас нет на это право' });
+    }
+
     const section = await Section.create({ board: boardId })
     section._doc.tasks = []
     res.status(201).json(section)
@@ -13,8 +22,17 @@ exports.create = async (req, res) => {
 }
 
 exports.update = async (req, res) => {
+  const { user } = req
   const { sectionId } = req.params
+
   try {
+    if (!user) {
+      throw new Error('User not found');
+    }
+    if (user.role !== 'Админ') {
+      return res.status(403).json({ error: 'У вас нет на это право' });
+    }
+
     const section = await Section.findByIdAndUpdate(
       sectionId,
       { $set: req.body }
@@ -27,8 +45,17 @@ exports.update = async (req, res) => {
 }
 
 exports.delete = async (req, res) => {
+  const { user } = req
   const { sectionId } = req.params
+
   try {
+    if (!user) {
+      throw new Error('User not found');
+    }
+    if (user.role !== 'Админ') {
+      return res.status(403).json({ error: 'У вас нет на это право' });
+    }
+
     await Task.deleteMany({ section: sectionId })
     await Section.deleteOne({ _id: sectionId })
     res.status(200).json('deleted')
